getters and setters tests made and all pass

+++ src/main/java/com/bae/service/FreezerService.java
@@ -20,8 +20,12 @@ public class FreezerService {
 	public Freezers createFreezer(Freezers freezer){
 		return this.repo.save(freezer);
 	}
-	public void  deleteFreezer(Long id) {
+	public boolean  deleteFreezer(Long id) throws FreezerDoesntexistException {
+		if(!this.repo.existsById(id)) {
+			throw new FreezerDoesntexistException();
+		}
 		this.repo.deleteById(id);
+		return this.repo.existsById(id);
 	}
